+++ lib/src/cmaf/ham/types/model/Segment.ts
@@ -1,0 +1,11 @@
+/**
+ * CMAF-HAM Segment type
+ *
+ * @group CMAF
+ * @alpha
+ */
+export type Segment = {
+	duration: number;
+	url: string;
+	byteRange?: string;
+};