+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.7.0] - 2024-06-05
 
 ### Added
+- Add CMAF-Ham types, mappers and services to convert and manipulate VOD HLS and VOD DASH manifests.
+- Add cmaf-ham-converter sample to showcase the CMAF-Ham functions.
 - Implement CEA 608/708 parser [#62](https://github.com/streaming-video-technology-alliance/common-media-library/issues/62)
 
 ### Fix